+++ CHANGES.md
@@ -6,7 +6,7 @@ All notable changes to this project will be documented in this file.
 
 - Adds `cookie` middleware.
 - New `Request::isMethod` and `Request::isJSON` and `Request::isAJAX`.
-- New `Request::getAccept`.
+- New `Request::getAccept`, `::getAcceptCharset`, `::getAcceptEncoding` and `::getAcceptLanguage`.
 - [BC-BREAK] Now `ServerRequest::getFiles` ignores empty files.
 
 ## [0.5.0] - 2019-02-21